fix: move `waits` -> `conditions`
