+++ sdks/python/examples/conditions/trigger.py
