+++ sdks/python/examples/conditions/worker.py
